+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Symbols are now quoted if they contain special characters.
+  - The symbol will be quote if any of the following characters in contained on
+    the symbol name: `@`, `<`, `\\`, `-` or `+`.
+
 ## [1.22.0] - 2024-03-08
 
 ### Added