+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `SectionRelocZ64` now emits `"a"` flag as the section flag.
+- Make sure to not reference local branch labels on data symbols.
 
 ## [1.22.0] - 2024-03-08