+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.23.0] - 2024-03-19
+
 ### Added
 
 - Symbols are now quoted if they contain special characters.
@@ -1445,6 +1447,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Version 1.0.0
 
 [unreleased]: https://github.com/Decompollaborate/spimdisasm/compare/master...develop
+[1.23.0]: https://github.com/Decompollaborate/spimdisasm/compare/1.22.0...1.23.0
 [1.22.0]: https://github.com/Decompollaborate/spimdisasm/compare/1.21.1...1.22.0
 [1.21.0]: https://github.com/Decompollaborate/spimdisasm/compare/1.20.1...1.21.0
 [1.20.1]: https://github.com/Decompollaborate/spimdisasm/compare/1.20.0...1.20.1